backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.22.1",
    "bcrypt<4",
    "easyocr>=1.7.2",
    "email-validator>=2.3.0",
    "fastapi>=0.128.0",
    "httpx>=0.28.1",
    "passlib[bcrypt]>=1.7.4",
    "pillow>=12.1.0",
    "protobuf>=6.33.5",
    "pydantic-settings>=2.12.0",
    "pymupdf>=1.26.7",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.22",
    "scikit-learn>=1.8.0",
    "sentencepiece>=0.2.1",
    "sqlalchemy>=2.0.46",
    "tiktoken>=0.12.0",
    "torch>=2.10.0",
    "transformers>=5.0.0",
    "uvicorn>=0.40.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
